Address review comments, move SQL DB configs to Java based.

+++ hadoop-ozone/recon/src/test/java/org/apache/hadoop/ozone/recon/persistence/AbstractReconSqlDBTest.java
@@ -210,12 +210,12 @@ public class AbstractReconSqlDBTest {
         }
 
         @Override
-        public Integer getMaxConnectionAge() {
+        public long getMaxConnectionAge() {
           return 120;
         }
 
         @Override
-        public Integer getMaxIdleConnectionAge() {
+        public long getMaxIdleConnectionAge() {
           return 120;
         }
 
@@ -225,7 +225,7 @@ public class AbstractReconSqlDBTest {
         }
 
         @Override
-        public Integer getIdleConnectionTestPeriod() {
+        public long getIdleConnectionTestPeriod() {
           return 30;
         }
       };
